pyproject.toml
[project]
name = "nonebot-plugin-mcp"
version = "0.1.0"
description = "NoneBot2 MCP Client Plugin - 提供 MCP 工具调用能力"
readme = "README.md"
requires-python = ">=3.10"

authors = [
    { name = "gsskk", email = "your.email@example.com" }
]
license = { file = "LICENSE" }

dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-plugin-alconna>=0.50.2",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.nonebot]
plugins = ["nonebot_plugin_mcp"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
ignore = ["E402"]
